Remove some common classes, spread helper functions to where they are used
Several classes in the common folder were only used by one library, so those classes have been moved to that library.

Some classes in the common folder had functions used only by one library and other functions used by a different library. Those functions have been split from the common class and moved to the library where they are used.

This is just a first step to removing the common folder, so we'll have more PRs like this in the future
