+++ iothub/device/src/Common/Amqp/ClientWebSocketTransport.cs
@@ -81,7 +81,7 @@ namespace Microsoft.Azure.Amqp.Transport
             Fx.AssertAndThrow(args.Buffer != null, "must have buffer to read");
             Fx.AssertAndThrow(args.CompletedCallback != null, "must have a valid callback");
 
-            Utils.ValidateBufferBounds(args.Buffer, args.Offset, args.Count);
+            ValidateBufferBounds(args.Buffer, args.Offset, args.Count);
             args.Exception = null; // null out any exceptions
 
             Task<int> taskResult = ReadImplAsync(args);
@@ -340,6 +340,25 @@ namespace Microsoft.Azure.Amqp.Transport
             }
 
             throw new AmqpException(AmqpErrorCode.IllegalState, null);
+        }
+
+        private static void ValidateBufferBounds(byte[] buffer, int offset, int size)
+        {
+            if (buffer == null)
+            {
+                throw new ArgumentNullException(nameof(buffer));
+            }
+
+            if (offset < 0 || offset > buffer.Length || size <= 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(offset));
+            }
+
+            int remainingBufferSpace = buffer.Length - offset;
+            if (size > remainingBufferSpace)
+            {
+                throw new ArgumentOutOfRangeException(nameof(size));
+            }
         }
     }
 }
