+++ iothub/service/src/AssertionFailedException.cs
