+++ iothub/service/src/Http/HttpMessageHelper.cs
