+++ iothub/service/src/Http/IHttpClientHelper.cs
