+++ iothub/service/src/Http/PutOperationType.cs
