+++ provisioning/service/src/Microsoft.Azure.Devices.Provisioning.Service.csproj
@@ -49,7 +49,6 @@
     <Compile Include="$(common)\service\CommonResources.cs" Link="Common\CommonResources.cs" />
     <Compile Include="$(common)\service\Resources.Designer.cs" Link="Common\Resources.Designer.cs" />
     <Compile Include="$(common)\service\StringValidationHelper.cs" Link="Common\StringValidationHelper.cs" />
-    <Compile Include="$(common)\service\Utils.cs" Link="Common\Utils.cs" />
     <Compile Include="$(Common)\DefaultWebProxySettings.cs" />
     <Compile Include="$(Common)\HttpContentExtensions.cs" Link="Utilities\HttpContentExtensions.cs" />
   </ItemGroup>
